Add specialist trait. Has to be changed a bit to work properly
The trait should be very useful to implement the different adventurers,
however, in it's current form, classes with extra drain or extra push
like engineer and navigator cannot be properly implemented, because
the character cannot ask them if they can currently actually drain a
position. This is why the structure has to be changed a little bit
still.